Update patch set 2
Patch Set 2: Verified+1

What are data we are getting as of now from backend . all that , i am populating here .
some of the data like -- utilization of cluster , etc it's not there yet. 

And as this time , we haven't backend data for populating in graph.
So once we have data , i will integrate with angular-patternfly-chart .

Patch-set: 2
Label: Verified=+1