fix blank name issue in rbd creation
In rbd creation page, if the user is increasing
the number of rbds created to more than and the name
is blank, the generated rbd names are wrong. This is
now fixed

Also if the number of rbds to be created is 1, then
index will not be appended to the name.

Change-Id: I2ebcf27a8c0ebc2c4c10bc8aa49104c474735cfc
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>